Cambiado precio a double

+++ app/src/main/java/edu/ucne/nathiel_taveras_ap2_p1/Data/Local/EntradasHuacalesEntity.kt
@@ -10,5 +10,5 @@ class EntradasHuacalesEntity(
     val Fecha: String = "",
     val NombreCliente: String = "",
     val Cantidad: Int,
-    val Precio: Float
+    val Precio: Double
 )
